# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizPortal Multi-Division API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/by-doc/{kind}/{id}:
    get:
      operationId: listPaymentsByDoc
      tags: [payments]
      summary: List payments for a document
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [sales, purchase] }
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /payments/sales/{id}/create-link:
    post:
      operationId: createSalesPaymentLink
      tags: [payments]
      summary: Create Paylabs payment link for a sales order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Payment link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLinkResponse"
        "202":
          description: Simulation mode (Paylabs not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLinkResponse"
  /payments/{id}/simulate-paid:
    post:
      operationId: simulatePaymentPaid
      tags: [payments]
      summary: Mark payment as paid (testing only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id: { type: integer }
        refKind: { type: string, enum: [sales, purchase] }
        refId: { type: integer }
        refDocNumber: { type: string }
        amount: { type: number }
        status: { type: string, enum: [pending, paid, expired, cancelled, failed] }
        provider: { type: string }
        providerOrderId: { type: string, nullable: true }
        providerMerchantTradeNo: { type: string }
        paymentUrl: { type: string, nullable: true }
        expiredAt: { type: string, nullable: true }
        paidAt: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, refKind, refId, refDocNumber, amount, status, provider, providerMerchantTradeNo, createdAt, updatedAt]
    PaymentLinkResponse:
      type: object
      properties:
        configured: { type: boolean }
        message: { type: string, nullable: true }
        payment: { $ref: "#/components/schemas/Payment" }
      required: [configured, payment]
...